// backend/src/domain/enums/contract-status.enum.ts
export enum ContractStatus {
  CREATED = "CREATED",
  SUPPLIER_REGISTRATION_OPENED = "SUPPLIER_REGISTRATION_OPENED",
  SUPPLIER_REGISTRATION_CLOSED = "SUPPLIER_REGISTRATION_CLOSED",

  DEPLOYED = "DEPLOYED",

  CUSTOMER_REGISTERED = "CUSTOMER_REGISTERED",
  SUPPLIERS_REGISTERED = "SUPPLIERS_REGISTERED",

  PENDING_PARTY_CONFIRMATION = "PENDING_PARTY_CONFIRMATION",
  DEMAND_CONFIRMED = "DEMAND_CONFIRMED",
  QUOTATIONS_CONFIRMED = "QUOTATIONS_CONFIRMED",

  CRITERIA_PENDING = "CRITERIA_PENDING",
  CRITERIA_SET = "CRITERIA_SET",
  ALLOCATED = "ALLOCATED",
  FUNDING = "FUNDING",
  FUNDED = "FUNDED",

  EXECUTING = "EXECUTING",
  COMPLETED = "COMPLETED",
}